# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/mails/all": {
      "get": {
        "operationId": "MailController_getMails",
        "summary": "Get all packages",
        "description": "Retrieve a paginated list of all packages with filtering options",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "receivedAtEnd",
            "required": false,
            "in": "query",
            "description": "Filter by received date end (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receivedAtStart",
            "required": false,
            "in": "query",
            "description": "Filter by received date start (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isForwarded",
            "required": false,
            "in": "query",
            "description": "Filter by forwarded status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isShereded",
            "required": false,
            "in": "query",
            "description": "Filter by shredded status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "carrier",
            "required": false,
            "in": "query",
            "description": "Filter by carrier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senderName",
            "required": false,
            "in": "query",
            "description": "Filter by sender name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "steNumber",
            "required": false,
            "in": "query",
            "description": "Filter by STE number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by package status",
            "schema": {
              "enum": [
                "PENDING",
                "FORWARDED",
                "SHREDDED",
                "COMPLETED",
                "CANCELLED",
                "IN_PROCESS"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by package type",
            "schema": {
              "enum": [
                "BANK_CHECK",
                "LEGAL_DOCUMENT",
                "ENVELOPE",
                "PACKAGE",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "officeLocationId",
            "required": false,
            "in": "query",
            "description": "Filter by office location ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceAddressId",
            "required": false,
            "in": "query",
            "description": "Filter by workspace address ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Packages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PackageResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Mail Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mails/{id}": {
      "get": {
        "operationId": "MailController_getMailById",
        "summary": "Get package by ID",
        "description": "Retrieve a specific package with all its items and related information",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Package ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Package found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Package not found"
          }
        },
        "tags": [
          "Mail Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Virtual Mail Hub",
    "description": "The Virtual Mail Hub API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PackageItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Item ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "packageId": {
            "type": "string",
            "description": "Package ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Item name",
            "example": "Legal Document"
          },
          "description": {
            "type": "string",
            "description": "Item description",
            "example": "Important legal document from law firm"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of items",
            "example": 1
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kilograms",
            "example": 0.5
          },
          "width": {
            "type": "number",
            "description": "Width in centimeters",
            "example": 21
          },
          "height": {
            "type": "number",
            "description": "Height in centimeters",
            "example": 29.7
          },
          "length": {
            "type": "number",
            "description": "Length in centimeters",
            "example": 0.1
          },
          "volumeDesi": {
            "type": "number",
            "description": "Volume in desi",
            "example": 0.1
          },
          "photoUrls": {
            "description": "Photo URLs",
            "example": [
              "https://s3.amazonaws.com/bucket/photo1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isShereded": {
            "type": "boolean",
            "description": "Is item shredded",
            "example": false
          },
          "isForwarded": {
            "type": "boolean",
            "description": "Is item forwarded",
            "example": false
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is item deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deletion date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "package": {
            "type": "object",
            "description": "Package details"
          }
        },
        "required": [
          "id",
          "packageId",
          "name",
          "quantity",
          "isShereded",
          "isForwarded",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "PackageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Package ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "steNumber": {
            "type": "string",
            "description": "STE number on the package",
            "example": "004712"
          },
          "workspaceAddressId": {
            "type": "string",
            "description": "Workspace address ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "officeLocationId": {
            "type": "string",
            "description": "Office location ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "type": {
            "type": "string",
            "description": "Package type",
            "enum": [
              "BANK_CHECK",
              "LEGAL_DOCUMENT",
              "ENVELOPE",
              "PACKAGE",
              "OTHER"
            ],
            "example": "PACKAGE"
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when package was received",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "senderName": {
            "type": "string",
            "description": "Sender name",
            "example": "John Doe Company"
          },
          "senderAddress": {
            "type": "string",
            "description": "Sender address",
            "example": "123 Main St, New York, NY 10001"
          },
          "carrier": {
            "type": "string",
            "description": "Carrier company",
            "example": "UPS"
          },
          "width": {
            "type": "number",
            "description": "Width in centimeters",
            "example": 30
          },
          "height": {
            "type": "number",
            "description": "Height in centimeters",
            "example": 20
          },
          "length": {
            "type": "number",
            "description": "Length in centimeters",
            "example": 15
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kilograms",
            "example": 2.5
          },
          "volumeDesi": {
            "type": "number",
            "description": "Volume in desi",
            "example": 9
          },
          "photoUrls": {
            "description": "Photo URLs",
            "example": [
              "https://s3.amazonaws.com/bucket/package1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Package status",
            "enum": [
              "PENDING",
              "FORWARDED",
              "SHREDDED",
              "COMPLETED",
              "CANCELLED",
              "IN_PROCESS"
            ],
            "example": "PENDING"
          },
          "isShereded": {
            "type": "boolean",
            "description": "Is package shredded",
            "example": false
          },
          "isForwarded": {
            "type": "boolean",
            "description": "Is package forwarded",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "items": {
            "description": "Package items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageItemResponseDto"
            }
          },
          "workspaceAddress": {
            "type": "object",
            "description": "Workspace address details"
          },
          "officeLocation": {
            "type": "object",
            "description": "Office location details"
          },
          "actions": {
            "description": "Package actions",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "forwardRequests": {
            "description": "Forward requests",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "steNumber",
          "workspaceAddressId",
          "officeLocationId",
          "type",
          "receivedAt",
          "status",
          "isShereded",
          "isForwarded",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}